@@ -1,11 +1,11 @@
-name: get_thumbnail_video
+name: video_thumbnail_pro
 description: "A flutter plugin for creating a thumbnail from a local video file or from a video URL."
 version: 0.7.3
-homepage:
-repository: https://github.com/Alberto-Monteiro/video_thumbnail
+homepage: https://github.com/arsarsars1/video_thumbnail_pro.git
+repository: https://github.com/arsarsars1/video_thumbnail_pro
 
 environment:
-  sdk: ^3.4.0
+  sdk: ^3.5.6
   flutter: ^3.22.0
 
 dependencies:
@@ -24,13 +24,13 @@ dev_dependencies:
 
 flutter:
   plugin:
-    implements: get_thumbnail_video
+    implements: video_thumbnail_pro
     platforms:
       android:
-        package: com.rocksti.get_thumbnail_video
+        package: com.ars.video_thumbnail_pro
         pluginClass: VideoThumbnailPlugin
       ios:
-        pluginClass: VideoThumbnailPlugin
+        pluginClass: VideoThumbnailProPlugin
       web:
-        pluginClass: VideoThumbnailWeb
+        pluginClass: VideoThumbnailProWeb
         fileName: video_thumbnail_web.dart
