@@ -1,21 +1,21 @@
-/// The Flutter plugin for creating thumbnail from video
-///
-/// To use, import `package:video_thumbnail/video_thumbnail.dart`.
-///
-/// See also:
-///
-///  * [video_thumbnail](https://pub.dev/packages/video_thumbnail)
-///
+// The Flutter plugin for creating thumbnail from video
+//
+// To use, import `package:video_thumbnail_pro/video_thumbnail_pro.dart`.
+//
+// See also:
+//
+//  * [video_thumbnail_pro](https://pub.dev/packages/video_thumbnail_pro)
+//
 import 'dart:async';
 
 import 'package:cross_file/cross_file.dart';
 import 'package:flutter/services.dart';
-import 'package:get_thumbnail_video/src/image_format.dart';
-import 'package:get_thumbnail_video/src/video_thumbnail_platform.dart';
+import 'package:video_thumbnail_pro/src/image_format.dart';
+import 'package:video_thumbnail_pro/src/video_thumbnail_platform.dart';
 
 export 'package:cross_file/cross_file.dart' show XFile;
 
-class VideoThumbnail {
+class VideoThumbnailPro {
   /// Generates a thumbnail file under specified thumbnail folder or given full path and name which matches expected ext.
   /// The video can be a local video file, or an URL repreents iOS or Android native supported video format.
   /// If the thumbnailPath is ommited or null, a thumbnail image file will be created under the same folder as the video file.
@@ -33,7 +33,7 @@ class VideoThumbnail {
   }) async {
     assert(video.isNotEmpty);
 
-    return VideoThumbnailPlatform.instance.thumbnailFile(
+    return VideoThumbnailProPlatform.instance.thumbnailFile(
       video: video,
       headers: headers,
       thumbnailPath: thumbnailPath,
@@ -60,7 +60,7 @@ class VideoThumbnail {
   }) async {
     assert(video.isNotEmpty);
 
-    return VideoThumbnailPlatform.instance.thumbnailData(
+    return VideoThumbnailProPlatform.instance.thumbnailData(
       video: video,
       headers: headers,
       imageFormat: imageFormat,
